updated meter reads and relationship with helper functions in the service and controller

+++ src/main/java/io/github/lordship/meterbills/MeterBills.java
@@ -10,7 +10,7 @@ public record MeterBills(
         UUID uuid,
         UUID billedMeter,
         Integer billedAmount,
-        Integer rateAmount,
+        Double rateAmount,
         MeterMeasurement rateUnit,
         LocalDate periodStart,
         LocalDate periodEnd,
